+++ curator/cli_singletons/allocation.py
@@ -10,9 +10,10 @@ from curator.cli_singletons.utils import get_width, validate_filter_json
 @click.option('--max_wait', default=-1, type=int, help='Maximum number of seconds to wait_for_completion', show_default=True)
 @click.option('--wait_interval', default=9, type=int, help='Seconds to wait between completion checks.', show_default=True)
 @click.option('--ignore_empty_list', is_flag=True, help='Do not raise exception if there are no actionable indices')
+@click.option('--allow_ilm_indices/--no-allow_ilm_indices', help='Allow Curator to operate on Index Lifecycle Management monitored indices.', default=False, show_default=True)
 @click.option('--filter_list', callback=validate_filter_json, help='JSON array of filters selecting indices to act on.', required=True)
 @click.pass_context
-def allocation(ctx, key, value, allocation_type, wait_for_completion, max_wait, wait_interval, ignore_empty_list, filter_list):
+def allocation(ctx, key, value, allocation_type, wait_for_completion, max_wait, wait_interval, ignore_empty_list, allow_ilm_indices, filter_list):
     """
     Shard Routing Allocation
     """
@@ -23,6 +24,7 @@ def allocation(ctx, key, value, allocation_type, wait_for_completion, max_wait, 
         'wait_for_completion': wait_for_completion,
         'max_wait': max_wait,
         'wait_interval': wait_interval,
+        'allow_ilm_indices': allow_ilm_indices,
     }
     # ctx.info_name is the name of the function or name specified in @click.command decorator
     action = cli_action(ctx.info_name, ctx.obj['config']['client'], manual_options, filter_list, ignore_empty_list)